fix evlr, has 8 bytes for recordlength instead of 2

+++ src/las/Vlr.cs
@@ -5,7 +5,7 @@ public class Vlr
     public ushort Reserved { get; set; }
     public string UserId { get; set; }
     public ushort RecordId { get; set; }
-    public ushort RecordLength { get; set; }
+    public long RecordLength { get; set; }
     public string Description { get; set; }
 
     public object Data { get; set; }
